Improving Blu-ray.com to list movies instead of releases
This way the website lists the movies it finds instead of the releases. Here the min year could be improved by having a variable with the year IMDB has an subtracting 1 or 2,  as there have been cases where Blu-ray.com listed a movie at certain year (ie. 2002) and IMDB has it with another year (ie. 2003) but in my experience those are rare cases.